+++ backend/.gitignore
@@ -16,3 +16,4 @@ data/jobs.csv
 data/meetOver.model
 data/rawTestUsers.json
 data/transcripts.csv
+matching/main.go
