+++ backend/data/stopwords.txt
@@ -1,41 +1,4 @@
-!!
-?!
-??
-!?
-`
-``
-''
--lrb-
--rrb-
--lsb-
--rsb-
-,
-.
-:
-;
-"
-'
-?
-<
->
-{
-}
-[
-]
-+
--
-(
-)
-&
-%
-$
-@
-!
-^
-#
-*
-..
-...
+###
 'll
 's
 'm
@@ -214,7 +177,6 @@ your
 yours
 yourself
 yourselves
-###
 return
 arent
 cant
